feat: :rocket: UserName. Finished Issue and made features.
Replace Public or Private with new text. In the case of Public and User being the owner the text becomes "Your Public Repository".  In the case of Private and User being the owner the text becomes "Your Private Repository".  In the case of Public and User being able to modify the text becomes "Team Public Repository".  In the case of Private and User being able to modify the text becomes "Team Private Repository".

Issues. https://github.com/VEERT00X/Github-Dev-Plus/issues/2.